front - Move DiscussionTabContent to dedicated component folder
