+++ front/src/app/components/discussion/DiscussionTabContent.tsx
